- Извлечен inline CSS из 11 админских шаблонов в отдельные .css файлы - Извлечен inline JavaScript из 11 админских шаблонов в отдельные .js файлы - Обновлены HTML файлы для подключения внешних CSS и JS через Thymeleaf - Добавлены улучшенные CSS классы для лучшего стилизования - Сохранена вся функциональность без изменений логики
Обработанные файлы:
- admin-dashboard.html → admin-dashboard.css + admin-dashboard.js
- admin-users.html → admin-users.css + admin-users.js
- admin-products-management.html → admin-products-management.css + admin-products-management.js
- admin-orders-management.html → admin-orders-management.css + admin-orders-management.js
- admin-add-product.html → admin-add-product.css + admin-add-product.js
- admin-add-brand.html → admin-add-brand.css + admin-add-brand.js
- admin-reviews-moderation.html → admin-reviews-moderation.css + admin-reviews-moderation.js
- admin-active-orders.html → admin-active-orders.css + admin-active-orders.js
- admin-archived-orders.html → admin-archived-orders.css + admin-archived-orders.js
- admin-active-order-details.html → admin-order-details.css + admin-active-order-details.js
- admin-archived-order-details.html → admin-order-details.css + admin-archived-order-details.js

+++ src/main/java/com/byteforge/byteforge/services/OrderService.java
@@ -92,7 +92,7 @@ public class OrderService {
                 order.getFirstName(),
                 savedOrder.getId(),
                 productNames,
-                totalPrice.doubleValue()
+                totalPrice
         );
 
         // 9. Возвращаем DTO
